Ensure feature works when element was constructed in script
This is a workaround test case for src/preload not working for _elements
upgraded before parsed_. The cause is that at the time of
`attributeChangedCallback`, element does not have access to
parentElemet as they haven't been fully parsed and is not yet connected
to the DOM.

The parsing order is difficult to re-create in test, and even harder to
make explicit, so this test reproduce the issue by creating "a upgraded
details-dialog, not connected to the DOM at the time of attribute
setting".

Note:

`attributeChangedCallback` always happens before `connectedCallback`,
therefore this test will pass if the attribute is set after
`connectedCallback`, triggering another `attributeChangedCallback` which
binds the events.